Python/controller/requirements.txt
python-binance==1.0.19
python-telegram-bot==12.8
